// src/components/pages/Profile/wallet/transactions/transactionDetail/types/transactionDetail.ts
export interface TransactionDetailLabels {
  title: string;
  purchaseInvoice: string;
  statusSuccess: string;
  trackingNumberLabel: string;
  trackingNumberValue: string;
  orderNumberLabel: string;
  orderNumberValue: string;
  viewOrder: string;
  transactionTypeLabel: string;
  transactionTypeValue: string;
  dateTimeLabel: string;
  dateTimeValue: string;
  orderAmountLabel: string;
  supportTitle: string;
  supportAction: string;
  supportIconAlt: string;
}